requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
loguru==0.7.2
python-docx==0.8.11
mammoth==1.6.0
pdfplumber==0.9.0
python-pptx==0.6.23
pandas==2.1.4
tabulate==0.9.0
markdownify==0.11.6
mistune==3.0.2
pillow==10.1.0
pytesseract==0.3.10
openai==1.6.1
redis==5.0.1
slowapi==0.1.9
openpyxl==3.1.2
aiofiles==23.2.0
chardet==5.2.0
requests==2.31.0 